Workaround being in a different directory
Before, `git diff` might be using the `relative` option, which would
only show changes in the current directory. This disables that option to
ensure we see *all* changed files, then ensures we are adding from the
correct directory (since the output of `git diff` will now be relative
to the repo root).