Allow 2 lines of Ghostty to show in accordioned windows
Ghostty's padding means that if we don't want the last line to be
cramped up against the bottom of the monitor, we have to have a whole
row of space. This measn we need to show a little more of the window to
see both lines of output we want for monitoring tasks.